Move tuple.h into 'internal' folder
We only ever use it internally, so we might as well mark it as such
