+++ src/adiar/bdd/bdd.h
@@ -2,10 +2,10 @@
 #define ADIAR_BDD_BDD_H
 
 #include <adiar/data.h>
-#include <adiar/tuple.h>
+#include <adiar/file.h>
 
 #include <adiar/internal/decision_diagram.h>
-#include <adiar/file.h>
+#include <adiar/internal/tuple.h>
 
 namespace adiar {
   class bdd;
