+++ src/adiar/bdd/if_then_else.cpp
@@ -2,9 +2,9 @@
 
 #include <adiar/file_stream.h>
 #include <adiar/file_writer.h>
-#include <adiar/tuple.h>
 
 #include <adiar/internal/levelized_priority_queue.h>
+#include <adiar/internal/tuple.h>
 #include <adiar/internal/util.h>
 
 #include <adiar/assert.h>
