+++ src/adiar/internal/quantify.h
@@ -4,13 +4,13 @@
 #include <variant>
 
 #include <adiar/data.h>
-#include <adiar/tuple.h>
 
 #include <adiar/file.h>
 #include <adiar/file_stream.h>
 #include <adiar/file_writer.h>
 
 #include <adiar/internal/levelized_priority_queue.h>
+#include <adiar/internal/tuple.h>
 
 namespace adiar
 {
